[Portal] Migration to unstyled
