--- docs/pages/api-docs/portal-unstyled.json
@@ -1,15 +1,0 @@
-{
-  "props": {
-    "children": { "type": { "name": "node" } },
-    "container": { "type": { "name": "union", "description": "HTML element<br>&#124;&nbsp;func" } },
-    "disablePortal": { "type": { "name": "bool" } }
-  },
-  "name": "PortalUnstyled",
-  "styles": { "classes": [], "globalClasses": {}, "name": null },
-  "spread": false,
-  "filename": "/packages/material-ui-unstyled/src/PortalUnstyled/PortalUnstyled.js",
-  "inheritance": null,
-  "demos": "",
-  "styledComponent": true,
-  "cssComponent": false
-}
