+++ docs/pages/api-docs/portal.json
@@ -5,9 +5,9 @@
     "disablePortal": { "type": { "name": "bool" } }
   },
   "name": "Portal",
-  "styles": { "classes": [], "globalClasses": {}, "name": "MuiPortal" },
+  "styles": { "classes": [], "globalClasses": {}, "name": null },
   "spread": false,
-  "filename": "/packages/material-ui/src/Portal/Portal.js",
+  "filename": "/packages/material-ui-unstyled/src/Portal/Portal.js",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/components/portal/\">Portal</a></li></ul>",
   "styledComponent": true,
