+++ packages/material-ui-unstyled/src/Portal/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './Portal';
+export * from './Portal';
