+++ packages/material-ui-unstyled/src/Portal/index.js
@@ -1,0 +1,1 @@
+export { default } from './Portal';
