--- packages/material-ui-unstyled/src/PortalUnstyled/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './PortalUnstyled';
-export * from './PortalUnstyled';
