--- packages/material-ui-unstyled/src/PortalUnstyled/index.js
@@ -1,1 +1,0 @@
-export { default } from './PortalUnstyled';
