+++ packages/material-ui-unstyled/src/index.js
@@ -4,7 +4,7 @@ export * from './BadgeUnstyled';
 export { default as SliderUnstyled } from './SliderUnstyled';
 export * from './SliderUnstyled';
 
-export { default as PortalUnstyled } from './PortalUnstyled';
+export { default as Portal } from './Portal';
 
 export { default as generateUtilityClass } from './generateUtilityClass';
