--- packages/material-ui/src/Portal/Portal.d.ts
@@ -1,17 +1,0 @@
-import { PortalUnstyledProps } from '@material-ui/unstyled';
-
-export type PortalProps = PortalUnstyledProps;
-
-/**
- * Portals provide a first-class way to render children into a DOM node
- * that exists outside the DOM hierarchy of the parent component.
- *
- * Demos:
- *
- * - [Portal](https://material-ui.com/components/portal/)
- *
- * API:
- *
- * - [Portal API](https://material-ui.com/api/portal/)
- */
-export default function Portal(props: PortalProps): JSX.Element;
