--- packages/material-ui/src/Portal/Portal.js
@@ -1,46 +1,0 @@
-import * as React from 'react';
-import PropTypes from 'prop-types';
-import { PortalUnstyled } from '@material-ui/unstyled';
-import { exactProp, HTMLElementType } from '@material-ui/utils';
-
-/**
- * Portals provide a first-class way to render children into a DOM node
- * that exists outside the DOM hierarchy of the parent component.
- */
-const Portal = React.forwardRef(function Portal(props, ref) {
-  return <PortalUnstyled {...props} ref={ref} />;
-});
-
-Portal.propTypes = {
-  // ----------------------------- Warning --------------------------------
-  // | These PropTypes are generated from the TypeScript type definitions |
-  // |     To update them edit the d.ts file and run "yarn proptypes"     |
-  // ----------------------------------------------------------------------
-  /**
-   * The children to render into the `container`.
-   */
-  children: PropTypes.node,
-  /**
-   * An HTML element or function that returns one.
-   * The `container` will have the portal children appended to it.
-   *
-   * By default, it uses the body of the top-level document object,
-   * so it's simply `document.body` most of the time.
-   */
-  container: PropTypes /* @typescript-to-proptypes-ignore */.oneOfType([
-    HTMLElementType,
-    PropTypes.func,
-  ]),
-  /**
-   * The `children` will be under the DOM hierarchy of the parent component.
-   * @default false
-   */
-  disablePortal: PropTypes.bool,
-};
-
-if (process.env.NODE_ENV !== 'production') {
-  // eslint-disable-next-line
-  Portal['propTypes' + ''] = exactProp(Portal.propTypes);
-}
-
-export default Portal;
