--- packages/material-ui/src/Portal/Portal.test.js
@@ -1,1 +1,0 @@
-describe('<Portal />', () => {});
