+++ packages/material-ui/src/Portal/index.js
@@ -1,1 +1,3 @@
-export { default } from './Portal';
+import { Portal } from '@material-ui/unstyled';
+
+export default Portal;
